Start actions and animations
This is a messy commit as I have not committed for way to long.
To manage the current action and state of this action I introduced
an ActionContext where Reducers can be added for every action. So
far, only the playCardReducer exists.
Additionally, I tried if animations are possible. As they need the
current and target position of the card, this position information
is stored in a map in the AnimationContext. The animation is
realized by a copy of the card, the FlyingCard.